+++ mondrian/src/main/java/mondrian/olap/Util.java
@@ -2178,7 +2178,7 @@ public class Util extends XOMUtil {
                 }
                 e1 = it1.next();
                 e2 = it2.next();
-            } else if (compare == 1) {
+            } else if (compare > 0) {
                 if (!it2.hasNext()) {
                     break;
                 }
