+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/AbstractAggregation.java
@@ -1,0 +1,43 @@
+package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlType;
+
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "AbstractAggregation", propOrder = {
+
+})
+public class AbstractAggregation {
+    @XmlElement(name = "ID")
+    protected String id;
+    @XmlElement(name = "Name", required = true)
+    protected String name;
+    @XmlElement(name = "Description")
+    protected String description;
+
+    public String getID() {
+        return id;
+    }
+
+    public void setID(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}
