+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/AbstractSession.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   SmartCity Jena - initial
+ *   Stefan Bischof (bipolis.org) - initial
+ */
+package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlAttribute;
+import jakarta.xml.bind.annotation.XmlSeeAlso;
+import jakarta.xml.bind.annotation.XmlType;
+
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "AbstractSession")
+@XmlSeeAlso({Session.class, EndSession.class})
+public class AbstractSession {
+
+    @XmlAttribute(name = "SessionId")
+    protected String sessionId;
+    @XmlAttribute(name = "mustUnderstand")
+    protected Integer mustUnderstand;
+
+    public String getSessionId() {
+        return sessionId;
+    }
+
+    public void setSessionId(String value) {
+        this.sessionId = value;
+    }
+
+    public int getMustUnderstand() {
+        return mustUnderstand;
+    }
+
+    public void setMustUnderstand(int value) {
+        this.mustUnderstand = value;
+    }
+
+}
